refactor(dashboard): component card
