+++ src/modules/dashboard/Board/components/CardBoard/components/CardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CardHeader";
