+++ src/modules/dashboard/Board/components/CardBoard/components/CardHeader/style/index.ts
@@ -1,0 +1,2 @@
+export * as default from "./CardHeader.style";
+export * from "./CardHeader.style";
