+++ src/modules/dashboard/Board/components/CardBoard/components/CardHeader/types/CardHeader.types.ts
@@ -1,0 +1,7 @@
+import { TypeBoard } from "shared/common/types/AppTypes";
+
+export interface ICardHeader {
+	type_board: TypeBoard;
+	board_id: string;
+	isFetching: boolean;
+}
