+++ src/modules/dashboard/Board/components/CardBoard/components/CardHeader/types/index.ts
@@ -1,0 +1,2 @@
+export * as default from "./CardHeader.types";
+export * from "./CardHeader.types";
