+++ src/modules/dashboard/Board/components/CardBoard/style/index.ts
@@ -1,5 +1,5 @@
 import {
-	Menu as MenuMui,
+	
 	CardContent as CardContentMui,
 	styled,
 } from "@mui/material/";
@@ -27,21 +27,6 @@ export const CardBoardContainer = styled("div")<ICardBoardStyle>(
 `
 );
 
-export const CardIcon = styled("div")(
-	({ theme }) => `
-        display: flex;
-        align-item: center;
-        gap: ${theme.spacing(1)}; 
-        text-transform: capitalize
-    `
-);
-
-export const CardHeader = styled("div")({
-	display: "flex",
-	alignItem: "center",
-	justifyContent: "space-between",
-});
-
 export const CardContent = styled(CardContentMui)(
 	() => `
         a {
@@ -56,23 +41,4 @@ export const CardContent = styled(CardContentMui)(
     `
 );
 
-export const Menu = styled(MenuMui)(
-	({ theme }) => `    
-    
-    .MuiPaper-root {
-        width: ${theme.spacing(15)};
-        border-radius: ${theme.spacing(1)};
-        box-shadow: none;
-        background: ${theme.palette.background.paper};
-    }
 
-        ul > li {
-            font-size: ${theme.typography.caption.fontSize};
-            padding-bottom: ${theme.spacing(0.5)};
-
-            > svg {
-                margin-right: ${theme.spacing(1)};
-            }
-        }
-    `
-);
