+++ src/shared/common/types/AppTypes/app.types.ts
@@ -1,3 +1,3 @@
-export type TypeBoard = "texts" | "notes";
+export type TypeBoard = "texts" | "notes" | "kanban";
 
 export type FunctionReturnsVoid = () => void;
